Update custom-backend.rst
1. "Currently its focus is to compile dependently type functional language in a timely manner. Its main purpose is checking if the given program is correct in a dependently typed setting and generate code in form of a lambda-calculus like IR where higher-order functions are present."
In this context, does setting mean environment?

2. Could you explain this sentence, so i can make suggestions?
Although by its other intermediate languages loosen the abstractions and make friendlier for other code generation techniques.

3. "The ``ClosedTerm`` is a special ``Term`` where the list of the unbound variables is empty. This
technicality is not important for the code generation of the custom back-end as the back-end needs to
call the ``getCompileData`` function which produces the ``Compiler.Common.CompileData`` record."
could you include that ``ClosedTerm`` is not an argument, etc. of ``getCompileData``?

4. For some reason, the following 2 lists doesn't appear as it should in the preview and i don't know why:
The CompileData contains: - A main expression that will be the entry point for the program in CExp - A list of Core.CompileExpr.NamedDef - A list of lambda-lifted definitions Compiler.LambdaLift.LiftedDef - A list of Compiler.ANF.ANFDef - A list of Compiler.VMCode.VMDef definitions

These lists contain - Functions - Top-level data definitions - Runtime crashes which represent unfilled holes in idris programs - Foreign call constructs.

5. Could you explain this sentence, so i can make suggestions?
The implementor of the custom back-end should pick the closest phase of the abstraction that
the custom back-end support making the transformation from one of the intermediate representation
of Idris to the custom back-end.

6. I would make 2 sentences out of this, but not sure how:
How to represent these primitive types must be a well-founded design decision as it affects many
part of the code generation, such as conversion from the back-end values when FFI is involved,
big part of the data during the runtime is represented in these forms.

7. "As the code snippets shows
the ``%World`` must be used linearly, which is a strong guarantee for the runtime system."
For what? For the safety of the runtime system?

8. "The prf which is guaranteed to be erased during runtime, due to its 0 quantity."
Is there a better word for quantity in this context? eg. instances?

9. "Which also exposes a restriction on the FFI data types."
What restriction?

10. "Official backends represents primitive data types as boxed ones."
Is there any info on the unofficial backands that could be briefly mentioned?

11. "These definitions not always contain an actual function definitions, but sometimes top-level data creation, or crash instructions."
Could you explain this sentence, so i can make suggestions?

12. "Formally a ``(css : List String)``, ``(fargs : List CFType)``,
and ``(ret : CFType)``."
Is this an example or an exact definition?

13. "Using this information the custom back-end needs to generate code in the
host technology which could invoke the function call in the host technology, wrapping and
unwrapping the Idris values (which are represented as CFType) between the runtime for the Idris
in the host technology and the foreign function."
Could you explain this sentence, so i can make suggestions?

14. "The third approach for expression is the approach can be found in the VMDef."
IR expressions? Approach to do *what* with these IR expressions?

15. "This is because Idris keeps around all the arguments,
but the code generator for the alternatives removes the ones which are marked for deletion,
for such arguments the code generator of the custom back-end needs to remove the erased
arguments for the constructor implementation."
This could be split up to two sentences, but I don't understand the second half, so i cannot rephrase it.

16. "As erased arguments don't hold any
variables and the arities of alternatives and constructors won't match up." 
How does the backend resolve this matching problem?

17. "Other functionality, which is part of the prelude/base,
can be think of abstract types via external and foreign
functions around them."
What is prelude/base? It hasn't been mentioned yet.

18. "From integers the 64Bits and arbitrary precision ones are supported,
From Word8 to Word64 are supported."
Could you explain this sentence, so i can make suggestions?

19. "One of the possible reasons to implement a custom back-end for Idris is to generate code for
another technology which has many libraries, but it doesn't have strong type properties."
What does not have strong type properties? The another technology?

20. "The decision here is how to convert from these values to values of the functions written in the host language? Convert values when precision is not adequate? Or stop compilation if such distinction is detected? What to do with possibly null String values?"
Shouldn't you have to attempt to answer your own questions here?

21. Could you please help explaining this section?
"Although it is not clear what back-end should support. Tools from the Scheme back-end
are brought into the Idris world via external types and with there primitive operations
around them. This is a good practice and gives the community the ability to focus on
the implementation of a quick quick compiler for a dependently typed language.
One of these hidden features is the currency primitives. These are part of the
different libraries that could be part of the compiler or part of the
contribution package. If the threading model is different for the host technology
that the Idris inherits currently from the Scheme technology it could be a bigger
piece of work."

22. "How to represent the state of the World, which part we are interested in"
Do you mean that we are only interested in one part of %world?

23. "If the implementation uses boxed primitive types,
storing constructor values on the heap will be uniform, that helps implementing or
inheriting the GC." is this an example?